Add table procedures support for Lakehouse
Iceberg module has several table procedure.
Hive and DeltaLake modules have only the "optimize" table procedure.
The code for the "optimize" table procedure is identical in all 3 modules.
So, the code from Iceberg module is used.
We could redeclare the table procedure in the Lakehouse module, or maybe there could be a refactoring so that the identical code for the "optimize" table procedure is placed in a common place and reused from there.